Remove weave from all PyCBC code: commit 1 (#2946)
Remove weave from skymax

Remove weave from pycbc_multi_inspiral

Remove weave from pycbc_inference

Remove weave-related kludge

Remove weave from pycbc_inspiral

Add note about doc page

Remove weave mention

Remove weave from example ipynb

Remove weave module

Not sure what to remove in __init__

Added question to Josh

Slight update to FFTW_pruned

Moving out simd_threshold to C-code

Looks like I missed these FIXMEs

Hook up simd_threshold in setup.py

FINALLY got this to compile!!!

Including threshold_cpu

Start work on the SIMD filter stuff

Some typos in the events SIMD code

Add decompress and update requirements

Remove some old weave files ... Not sure where used

Fix compilation

Remove stuff from the bundler script

Remove a bunch of weave, not sure what this does

I'm not sure this is used, but would need updating

Some final touch-ups

Apparently the .h is bad in complex
