+++ bin/hdfcoinc/pycbc_calculate_psd
@@ -88,17 +88,6 @@ logging.info('%d psds to calculate', len(segments))
 
 if len(segments) > 0:
     pool = multiprocessing.Pool(args.cores)
-
-    # KLUDGE! Run the first segment to esure that the weave cache is
-    # populated.  This is a short-term fix until the issues with
-    # https://github.com/ligo-cbc/pycbc/issues/501
-    # can be resolved.  This will wastefully run the first segment
-    # twice which could be avoided, but since this is a short-term
-    # kludge anyway I'd prefer to be minimally invasive.
-    # TODO: Remove this when 501 is resolved.
-    # FIXME: https://www.youtube.com/watch?v=DtRNg5uSKQ0
-    get_psd( (segments[0], 0) )
-
     psds = pool.map_async(get_psd, zip(segments, range(len(segments))))
     psds = psds.get()
 else:
