+++ bin/pycbc_inspiral
@@ -27,7 +27,6 @@ from pycbc.filter import MatchedFilterControl, make_frequency_series, qtransform
 from pycbc.types import TimeSeries, FrequencySeries, zeros, float32, complex64
 import pycbc.version
 import pycbc.opt
-import pycbc.weave
 import pycbc.inject
 import time
 
@@ -194,7 +193,6 @@ strain.StrainSegments.insert_segment_option_group(parser)
 scheme.insert_processing_option_group(parser)
 fft.insert_fft_option_group(parser)
 pycbc.opt.insert_optimization_option_group(parser)
-pycbc.weave.insert_weave_option_group(parser)
 pycbc.inject.insert_injfilterrejector_option_group(parser)
 SingleDetSGChisq.insert_option_group(parser)
 opt = parser.parse_args()
@@ -206,7 +204,6 @@ strain.StrainSegments.verify_segment_options(opt, parser)
 scheme.verify_processing_options(opt, parser)
 fft.verify_fft_options(opt,parser)
 pycbc.opt.verify_optimization_options(opt, parser)
-pycbc.weave.verify_weave_options(opt, parser)
 
 pycbc.init_logging(opt.verbose)
