+++ bin/pycbc_inspiral_skymax
@@ -24,7 +24,6 @@ from pycbc.filter import MatchedFilterSkyMaxControl, make_frequency_series, Matc
 from pycbc.types import TimeSeries, FrequencySeries, zeros, float32, complex64
 import pycbc.fft.fftw
 import pycbc.opt
-import pycbc.weave
 
 parser = argparse.ArgumentParser(usage='',
     description="Find single detector precessing gravitational-wave triggers.")
@@ -152,7 +151,6 @@ strain.StrainSegments.insert_segment_option_group(parser)
 scheme.insert_processing_option_group(parser)
 fft.insert_fft_option_group(parser)
 pycbc.opt.insert_optimization_option_group(parser)
-pycbc.weave.insert_weave_option_group(parser)
 SingleDetSGChisq.insert_option_group(parser)
 
 opt = parser.parse_args()
@@ -164,7 +162,6 @@ strain.StrainSegments.verify_segment_options(opt, parser)
 scheme.verify_processing_options(opt, parser)
 fft.verify_fft_options(opt,parser)
 pycbc.opt.verify_optimization_options(opt, parser)
-pycbc.weave.verify_weave_options(opt, parser)
 
 pycbc.init_logging(opt.verbose)
