+++ requirements.txt
@@ -6,13 +6,11 @@ decorator>=3.4.2
 scipy>=0.16.0; python_version >= '3.5'
 scipy>=0.16.0,<1.3.0; python_version <= '3.4'
 matplotlib>=2.0.0
-numpy>=1.13.0,<1.15.3; python_version <= '2.7'
-numpy>=1.13.0; python_version >= '3.0'
+numpy>=1.13.0
 pillow
 h5py<2.10.0
 jinja2
 mpld3>=0.3
-weave>=0.16.0; python_version <= '2.7'
 requests>=1.2.1
 beautifulsoup4>=4.6.0
 six>1.10.0
