+++ setup.py
@@ -34,8 +34,7 @@ from setuptools.command.build_ext import build_ext as _build_ext
 from setuptools import find_packages
 
 requires = []
-setup_requires = ['numpy>=1.13.0,<1.15.3; python_version <= "2.7"',
-                  'numpy>=1.13.0; python_version > "3.0"']
+setup_requires = ['numpy>=1.13.0']
 install_requires =  setup_requires + ['Mako>=1.0.1',
                       'cython',
                       'decorator>=3.4.2',
@@ -55,7 +54,6 @@ install_requires =  setup_requires + ['Mako>=1.0.1',
                       'six>=1.10.0',
                       'ligo-segments',
                       'tqdm',
-                      'weave>=0.16.0; python_version <= "2.7"',
                       ]
 
 def find_files(dirname, relpath=None):
@@ -241,6 +239,29 @@ e = Extension("pycbc.events.eventmgr_cython",
               extra_link_args=cython_link_args,
               compiler_directives={'embedsignature': True})
 ext.append(e)
+e = Extension("pycbc.events.simd_threshold_cython",
+              ["pycbc/events/simd_threshold_cython.pyx"],
+              language='c++',
+              extra_compile_args=cython_compile_args,
+              extra_link_args=cython_link_args,
+              compiler_directives={'embedsignature': True})
+ext.append(e)
+e = Extension("pycbc.filter.simd_correlate_cython",
+              ["pycbc/filter/simd_correlate_cython.pyx"],
+              language='c++',
+              extra_compile_args=cython_compile_args,
+              extra_link_args=cython_link_args,
+              compiler_directives={'embedsignature': True})
+ext.append(e)
+e = Extension("pycbc.waveform.decompress_cpu_cython",
+              ["pycbc/waveform/decompress_cpu_cython.pyx"],
+              language='c++',
+              extra_compile_args=cython_compile_args,
+              extra_link_args=cython_link_args,
+              compiler_directives={'embedsignature': True})
+ext.append(e)
+
+
 
 
 setup (
